requirements.txt
aiogram==3.*
asyncpg==0.29.*
python-dotenv==1.*
googletrans==4.0.0-rc1
